prototype/requirements.txt
# GCP 클라이언트
google-cloud-billing==1.14.1
google-cloud-bigquery==3.20.0

# AWS 클라이언트
boto3==1.28.85
botocore==1.31.85

# LLM API
google-generativeai==0.3.0
anthropic==0.25.0
httpx==0.24.0

# 데이터 처리
pandas==2.1.3
numpy==1.24.3

# 이상 탐지
scipy==1.11.4

# 유틸리티
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1

# 개발용 (선택)
jupyter==1.0.0
pytest==7.4.3
black==23.12.0

# 데이터 다운로드
datasets